+++ trilby-cli/src/Trilby/Install/Flake.hs
@@ -1,8 +1,8 @@
 module Trilby.Install.Flake where
 
 import Data.Text qualified as Text
-import Trilby.Install.Config.Release (Release (..))
 import Trilby.HNix hiding (Flake)
+import Trilby.Install.Config.Release (Release (..))
 import Prelude
 
 data InputOverride = Follows Text Text