+++ trilby-cli/src/Trilby/Update.hs
@@ -1,21 +1,26 @@
 module Trilby.Update (update) where
 
 import Data.List.NonEmpty.Extra qualified as NonEmpty
-import Trilby.HNix (FileOrFlake (..), FlakeRef (..), nixBuild, writeNixFile)
-import Trilby.Update.Options
-import Turtle (readlink, (</>))
-import Trilby.Host
-import Trilby.Configuration (Configuration(..))
+import Trilby.Configuration (Configuration (..))
 import Trilby.Configuration qualified as Configuration
+import Trilby.HNix (FileOrFlake (..), FlakeRef (..), copyClosure, nixBuild, writeNixFile)
+import Trilby.Host
+import Trilby.Update.Options
+import Turtle (parent, readlink, (</>))
+import Turtle qualified
+import UnliftIO.Directory (canonicalizePath)
 import Prelude
 
 update :: UpdateOpts Maybe -> App ()
 update (askOpts -> opts) = do
-    whenM opts.flakeUpdate . inDir "/etc/trilby" $ rawCmd_ ["nix", "flake", "update", "--accept-flake-config"]
+    whenM opts.flakeUpdate do
+        trilbyHome <- canonicalizePath "/etc/trilby"
+        flakes <- Turtle.sort $ parent <$> Turtle.find (Turtle.suffix "/flake.lock") trilbyHome
+        mapM_ updateFlake flakes
     configurations <- mapM Configuration.fromHost . NonEmpty.nubOrd =<< opts.hosts
     configurationResults <- buildConfigurations configurations
     for_ configurationResults $ \(Configuration{..}, resultPath) -> do
-        unless (host == Localhost) $ copyClosure host resultPath
+        copyClosure host resultPath
         ssh host rawCmd_ ["unbuffer", "nvd", "diff", "/run/current-system", fromString resultPath]
         unless (host == Localhost && length configurationResults == 1) $ $(logWarn) $ "Choosing action for host " <> ishow host
         let perform = switchToConfiguration host resultPath
@@ -26,6 +31,9 @@ update (askOpts -> opts) = do
                 whenM reboot $ ssh host cmd_ ["systemctl", "reboot"]
             Test -> perform ConfigTest
             NoAction -> pure ()
+
+updateFlake :: FilePath -> App ()
+updateFlake path = rawCmd_ ["nix", "flake", "update", "--accept-flake-config", "--flake", fromString path]
 
 buildConfiguration :: Configuration -> App (Configuration, FilePath)
 buildConfiguration c = (c,) <$> nixBuild f
@@ -60,10 +68,6 @@ buildConfigurations configurations = withSystemTempFile "trilby-update-.nix" $ \
          |]
     resultPath <- nixBuild $ File tmpFile
     forM configurations $ \c -> (c,) <$> readlink (resultPath </> fromText c.name)
-
-copyClosure :: Host -> FilePath -> App ()
-copyClosure Localhost _ = pure ()
-copyClosure host@Host{} path = cmd_ ["nix-copy-closure", "--gzip", "--to", ishow host, fromString path]
 
 data ConfigAction
     = ConfigBoot